[change]viewsファイルを分割 [change]modelからjsonへのパースを関数化した
