+++ api/views.py
@@ -151,281 +151,6 @@ def merged_blocks_change_streaming(message, map_id):
     return
 
 
-@api_app.route('/get_blocks/<uuid:map_id>/')
-def get_blocks(map_id):
-    blocks = db.session.query(Block).filter_by(map_id=map_id)
-
-    data = {
-        "blocks": [],
-    }
-    for block in blocks:
-        data["blocks"].append({
-            "ID": block.id,
-            "colorID": block.colorID,
-            "time": block.time,
-            "x": block.x,
-            "y": block.y,
-            "z": block.z,
-            "pattern_name": block.pattern_name,
-            "pattern_group_id": str(block.pattern_group_id) if block.pattern_group_id else None
-        })
-
-    return make_response(jsonify(data))
-
-
-@api_app.route('/get_maps/')
-def get_maps():
-    maps = db.session.query(Map)
-    maps_data = {
-        "maps": []
-    }
-    for map in maps:
-        tmp_dic = {
-            "ID": map.id,
-            "name": map.name
-        }
-        maps_data["maps"].append(tmp_dic)
-
-    return make_response(jsonify(maps_data))
-
-
-@api_app.route('/create_map/', methods=["POST"])
-def create_map():
-    if request.content_type == "application/json":
-        try:
-            request.json["name"]
-            request.json["realsense_id"]
-        except KeyError:
-            return make_response('name or realsense missing'), 400
-
-        name = request.json["name"]
-        new_map = Map(name=name)
-        db.session.add(new_map)
-        try:
-            db.session.commit()
-        except:
-            db.session.rollback()
-            return make_response('integrity error'), 500
-
-        try:
-            realsense_id = request.json["realsense_id"]
-            realsense = db.session.query(RealSense).filter_by(id=realsense_id).first()
-            realsense.current_map_id = new_map.id
-            db.session.commit()
-        except:
-            db.session.rollback()
-            return make_response('integrity error'), 500
-
-        map_data = {
-            "ID": new_map.id,
-            "name": name
-        }
-        return make_response(jsonify(map_data))
-    else:
-        return make_response('content type must be application/app'), 406
-
-
-@api_app.route('/get_color_rules/<uuid:map_id>/')
-def get_color_rules(map_id):
-    color_rules = db.session.query(ColorRule).filter_by(map_id=map_id)
-    data = {
-        "rules": []
-    }
-    for rule in color_rules:
-        if rule.type == "ID":
-            data["rules"].append({
-                "type": rule.type,
-                "block_id": rule.block_id,
-                "to": rule.to,
-                "map_id": rule.map_id
-            })
-        else:
-            data["rules"].append({
-                "type": rule.type,
-                "origin": rule.origin,
-                "to": rule.to,
-                "map_id": rule.map_id
-            })
-    return make_response(jsonify(data))
-
-
-@api_app.route('/create_color_rule/', methods=["POST"])
-def create_color_rule():
-    if request.content_type == "application/json":
-        try:
-            request.json["map_id"]
-        except KeyError:
-            return make_response('map_id missing'), 400
-        try:
-            request.json["type"]
-        except KeyError:
-            return make_response('type missing'), 400
-        try:
-            request.json["to"]
-        except KeyError:
-            return make_response('to missing'), 400
-
-        type = request.json["type"]
-        if type == "ID":
-            try:
-                request.json["block_id"]
-            except KeyError:
-                return make_response('block_id missing'), 400
-
-            block_id = request.json["block_id"]
-            to = request.json["to"]
-            map_id = request.json["map_id"]
-            db.session.add(ColorRule(type=type, block_id=block_id, to=to, map_id=map_id))
-        else:
-            try:
-                request.json["origin"]
-            except KeyError:
-                return make_response('origin missing'), 400
-
-            origin = request.json["origin"]
-            to = request.json["to"]
-            map_id = request.json["map_id"]
-            db.session.add(ColorRule(type=type, origin=origin, to=to, map_id=map_id))
-
-        try:
-            db.session.commit()
-        except:
-            db.session.rollback()
-            return make_response('integrity error'), 500
-
-        return make_response("ok")
-    else:
-        return make_response('content type must be application/app'), 406
-
-
-@api_app.route('/get_merges/')
-def get_merges():
-    merges = db.session.query(Merge)
-    data = {
-        "merges": []
-    }
-    for merge in merges:
-        data["merges"].append({
-            "ID": merge.id,
-            "name": merge.name
-        })
-    return make_response(jsonify(data))
-
-
-@api_app.route('/create_merge/', methods=["POST"])
-def create_merge():
-    if request.content_type == "application/json":
-        try:
-            request.json["name"]
-            request.json["merge_maps"]
-        except KeyError:
-            return make_response('name or merge_maps missing'), 400
-
-        new_merge = Merge(name=request.json["name"])
-        db.session.add(new_merge)
-        try:
-            db.session.commit()
-        except:
-            db.session.rollback()
-            return make_response('integrity error'), 500
-
-        merge_id = new_merge.id
-        merge_maps = request.json["merge_maps"]
-        merge_map_objects = []
-        for merge_map_data in merge_maps:
-            try:
-                merge_map_data["map_id"]
-            except KeyError:
-                return make_response('map_id missing'), 400
-            try:
-                merge_map_data["x"]
-                merge_map_data["y"]
-            except KeyError:
-                return make_response('x or y missing'), 400
-            try:
-                merge_map_data["rotate"]
-            except KeyError:
-                return make_response('rotate missing'), 400
-
-            merge_map_objects.append(
-                MergeMap(
-                    map_id=merge_map_data["map_id"],
-                    merge_id=merge_id,
-                    x=merge_map_data["x"],
-                    y=merge_map_data["y"],
-                    rotate=merge_map_data["rotate"]
-                )
-            )
-        db.session.bulk_save_objects(merge_map_objects, return_defaults=True)
-        try:
-            db.session.commit()
-        except:
-            db.session.rollback()
-            return make_response('integrity error'), 500
-
-        return make_response('ok')
-    else:
-        return make_response('content type must be application/json'), 406
-
-
-
-@api_app.route('/get_merge_maps/<uuid:merge_id>/')
-def get_merge_maps(merge_id):
-    merge_maps = db.session.query(MergeMap).filter_by(merge_id=merge_id)
-    data = {
-        "merge_maps": []
-    }
-    for merge_map in merge_maps:
-        data["merge_maps"].append({
-            "map_id": merge_map.map_id,
-            "x": merge_map.x,
-            "y": merge_map.y,
-            "rotate": merge_map.rotate
-        })
-    return make_response(jsonify(data))
-
-
-@api_app.route('/get_merged_blocks/<uuid:merge_id>/')
-def get_merged_blocks(merge_id):
-    merge_maps = db.session.query(MergeMap).filter_by(merge_id=merge_id).all()
-    merged_blocks = []
-
-    for merge_map in merge_maps:
-        blocks = db.session.query(Block).filter_by(map_id=merge_map.map_id).all()
-        for _block in blocks:
-            """
-            ブロックの座標移動処理
-            """
-            block = copy.deepcopy(_block)
-            rad = radians(90 * merge_map.rotate)
-            tmp_x = block.x
-            tmp_z = block.z
-            block.x = round(tmp_x * cos(rad) - tmp_z * sin(rad))
-            block.z = round(tmp_z * cos(rad) + tmp_x * sin(rad))
-
-            block.x += merge_map.x
-            block.z += merge_map.y
-
-            merged_blocks.append(block)
-
-    data = {
-        "blocks": [],
-    }
-    for merged_block in merged_blocks:
-        data["blocks"].append({
-            "ID": merged_block.id,
-            "colorID": merged_block.colorID,
-            "time": merged_block.time,
-            "x": merged_block.x,
-            "y": merged_block.y,
-            "z": merged_block.z,
-            "pattern_name": merged_block.pattern_name,
-            "pattern_group_id": str(merged_block.pattern_group_id) if merged_block.pattern_group_id else None
-        })
-
-    return make_response(jsonify(data))
-
-
 def recognize_pattern(blocks, map_id):
     """
     patterns sample
@@ -686,19 +411,3 @@ def merge_found_object(pattern, found_pattern_objects, merge_base_object, merged
             merged_objects = merge_found_object(pattern, found_pattern_objects, can_merge_object, merged_objects)
 
     return merged_objects
-
-@api_app.route('/get_realsense/')
-def get_realsense_fn():
-    data = {}
-    data["realsense"] = []
-    for realsense in db.session.query(RealSense).all():
-        print(realsense)
-        current_map = ""
-        if (realsense.current_map_id is not None):
-            current_map = db.session.query(Map).filter_by(id=realsense.current_map_id).first().name
-        data["realsense"].append({
-            "name": realsense.name,
-            "online": True, # To-DO
-            "current_map": current_map
-        })
-    return make_response(jsonify(data))
